pyproject.toml
[tool.poetry]
name = "webapi"
version = "0.1.0"
description = ""
authors = ["joseporiolayats <oriol.ayats@gmail.com>"]
readme = "README.md"
packages = [{include = "webapi"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
pre-commit = "^3.3.1"
pymongo = "^4.3.3"
colorlog = "^6.7.0"
python-dotenv = "^1.0.0"
pytest = "^7.3.1"
pydantic = {extras = ["email"], version = "^1.10.7"}
python-decouple = "^3.8"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
motor = "^3.1.2"
aiohttp = "^3.8.4"
email-validator = "^2.0.0.post2"
pyjwt = "^2.6.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
aiocache = "^0.12.1"


[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.3"
mkdocs-material = "^9.1.9"
mkdocstrings-python = "^0.9.0"


[tool.poetry.group.testing.dependencies]
pytest = "^7.3.1"
ruff = "^0.0.264"


[tool.poetry.group.lint.dependencies]
black = "^23.3.0"
ruff = "^0.0.264"


[tool.poetry.group.test.dependencies]
httpie = "^3.2.1"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
